rename Steps directory to SignInSteps and move it to components
