+++ src/components/index.ts
@@ -23,3 +23,4 @@ export * from './Dialogs'
 export { default as StudioContainer } from './StudioContainer'
 export * from './Topbar'
 export { default as NoConnectionIndicator } from './NoConnectionIndicator'
+export * from './SignInSteps'
